Moved homepage logic to homepage.ts, only function calls remain in main
